src/mcp_ml/requirements.txt
tensorflow>=2.0.0
numpy>=1.19.2
pandas>=1.1.3
matplotlib>=3.3.2
pillow>=8.0.1
scikit-learn>=0.23.2
pyyaml>=5.3.1
